+++ eedc/frontend/src/components/discovery/index.ts
@@ -1,8 +1,0 @@
-/**
- * Discovery Components Module
- */
-
-export { default as DiscoveryDialog } from './DiscoveryDialog'
-export { default as DeviceCard } from './DeviceCard'
-export { default as SensorMappingPanel } from './SensorMappingPanel'
-export { default as ConfirmationSummary } from './ConfirmationSummary'