+++ eedc/frontend/src/components/setup-wizard/index.ts
@@ -1,6 +1,6 @@
 /**
  * Setup-Wizard Komponenten Export
- * v0.8.0 - Refactored
+ * v1.1.0 - Bereinigt (Discovery entfernt)
  */
 
 export { default as SetupWizard } from './SetupWizard'
@@ -10,8 +10,6 @@ export { default as WelcomeStep } from './steps/WelcomeStep'
 export { default as AnlageStep } from './steps/AnlageStep'
 export { default as HAConnectionStep } from './steps/HAConnectionStep'
 export { default as StrompreiseStep } from './steps/StrompreiseStep'
-export { default as DiscoveryStep } from './steps/DiscoveryStep'
 export { default as InvestitionenStep } from './steps/InvestitionenStep'
-export { default as SensorConfigStep } from './steps/SensorConfigStep'
 export { default as SummaryStep } from './steps/SummaryStep'
 export { default as CompleteStep } from './steps/CompleteStep'